# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MédiGuide API",
    "description": "API de localisation de professionnels de santé (médecins, pharmacies, parapharmacies) à Jendouba, Tunisie : recherche géographique, prise de rendez-vous, avis, espace professionnel et administration. Un module complémentaire (medieye) expose ces mêmes fiches au format Facility générique de la plateforme partagée mediEYE/craftEYE/cityEYE sous /api/modules/medieye — voir openapi-medieye-fragment.json.",
    "version": "1.0.0",
    "contact": {
      "name": "MédiGuide"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Développement local"
    },
    {
      "url": "https://<votre-projet>.up.railway.app/api",
      "description": "Production (Railway)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT d'accès posé en cookie httpOnly après /api/auth/login"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Requis sur toute requête POST/PATCH/DELETE ; valeur du cookie csrf_token obtenu via GET /api/auth/csrf"
      }
    },
    "schemas": {
...
      "Disponibilite": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "facilityId": {
            "type": "integer"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "heureDebut": {
            "type": "string",
            "example": "09:00"
          },
          "heureFin": {
            "type": "string",
            "example": "09:20"
          },
          "estReserve": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/facilities/{facilityId}/disponibilites": {
      "get": {
        "tags": [
          "Disponibilités"
        ],
        "summary": "Créneaux libres et futurs d'un établissement",
        "security": [],
        "parameters": [
          {
            "name": "facilityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Créneaux disponibles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Disponibilite"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/disponibilites/mine": {
      "get": {
        "tags": [
          "Disponibilités"
        ],
        "summary": "Tous les créneaux du professionnel connecté (avec patient si réservé)",
        "responses": {
          "200": {
            "description": "Créneaux"
          },
          "403": {
            "description": "Rôle professionnel requis"
          }
        }
      }
    },
    "/disponibilites": {
      "post": {
        "tags": [
          "Disponibilités"
        ],
        "summary": "Générer un lot de créneaux",
        "description": "Génère des créneaux sur une plage horaire + durée ; ignore automatiquement les chevauchements avec des créneaux existants.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "facilityId",
                  "date",
                  "heureDebut",
                  "heureFin",
                  "dureeMinutes"
                ],
                "properties": {
                  "facilityId": {
                    "type": "integer"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "heureDebut": {
                    "type": "string"
                  },
                  "heureFin": {
                    "type": "string"
                  },
                  "dureeMinutes": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Créneaux créés"
          },
          "403": {
            "description": "Rôle professionnel requis"
          }
        }
      }
    },
    "/disponibilites/{id}": {
      "delete": {
        "tags": [
          "Disponibilités"
        ],
        "summary": "Supprimer un créneau non réservé",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimé"
          },
          "403": {
            "description": "Propriétaire uniquement"
          },
          "409": {
            "description": "Créneau déjà réservé"
          }
        }
      }
    },
...
  }
}